refactor(auth): simplify login and OTP pages by delegating to separate components and home
