+++ frontend-customer/src/app/(main)/page.tsx
@@ -1,5 +1,4 @@
-'use client'
-import HomePage from './(page)/home/HomePage'
+import HomePage from '@/views/Home'
 
 export default function Home() {
   return (
